requirement.txt
ultralytics
opencv-python
Pillow
numpy
ipython